+++ app/Http/Controllers/DashboardController.php
@@ -45,6 +45,7 @@ class DashboardController extends Controller
                             'overall_experience' => $stat->skills['Overall']['experience'] ?? 0,
                             'overall_level' => $stat->skills['Overall']['level'] ?? 0,
                             'skills' => $stat->skills ?? [],
+                            'activities' => $stat->activities ?? [],
                         ];
                     })
                     ->toArray();